+++ ee/api/entrypoint.sh
@@ -2,4 +2,4 @@
 sh env_vars.sh
 source /tmp/.env.override
 
-uvicorn app:app --host 0.0.0.0 --port $LISTEN_PORT --reload --proxy-headers
+uvicorn app:app --host 0.0.0.0 --port $LISTEN_PORT --proxy-headers